I add basic authentication to the children part of the website.
I split the main dispatch handler into two, one for main pages and one for the children pages.
The children handler is then wrapped in an authentication function which checks a cryptographically-encoded
password retrieved from the DB against the one the user passed in. The user can only access the children
pages if the check is successful.